requirements.txt
mysqlclient
python-dotenv
djangorestframework
djangorestframework-simplejwt
stripe
Pillow
pandas
pytz
twilio
pytest-mock
python-dateutil